+++ CHANGELOG.md
@@ -15,6 +15,16 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Breaking changes
+
+- Remove `doc` script.
+
+### Enhancements
+
+- Add a `readme` script to generate a README file from a template.
+
 ## 0.17.0 (November 9, 2018)
 
 ### Breaking changes